TEST: sprite creation

+++ tests/create-sprite.sikuli/create-sprite.py
@@ -12,7 +12,7 @@ bsfn= os.path.splitext(os.path.basename(fn))[0]
 addImagePath(path)
 
 def main():
-    spriteFName= bsfn + ".spr"
+    spriteFName= "GG100.txt"
     wait("ctp2start-scr.png", 100) # still works with sprite button shown
     click("ctp2sprite-test-btn.png")
     wait(10) # progress bar does not appear at once
@@ -21,12 +21,17 @@ def main():
         click(Pattern("ctp2sprite-sprite-name-field.png").similar(0.95).targetOffset(0, -10), 100) # text field needs click to get focus
         Settings.TypeDelay = 0.1; # ctp2-SDL needs some more time
         type(spriteFName)
-        click("ctp2sprite-save-btn.png")
+        click("ctp2sprite-load-btn.png")
         wait(10)
-        file = capture(SCREEN.getBounds())
-        if file:
-            shutil.move(file, bsfn + '.png')
-            exit(0)
+        click("ctp2sprite-save-btn.png")
+        if exists("ctp2sprite-mushroom-start-img.png", 100):
+            wait(10)
+            file = capture(SCREEN.getBounds())
+            if file:
+                shutil.move(file, bsfn + '.png')
+                exit(0)
+        else:
+            exit(30)
     else:
         exit(20)
     exit(99)
